Rearrange modules and introduce qmake project files.
After discussion with Justin, we decided to have two layers of modules.
This adds a little complication in the build process, but not too much.
Currently we just have one module, which is core->common, which is where
I've stuck all the basic classes for now.

I also added qmake project files so we should be able to build the
core->common module and also test it.