MOS-53 add virtual constructor to base class
I added a virtual destructor to the isx:Object base class, which is
currently used by isx::Ratio and isx::Time.

The Object destructor does not do anything and neither of those sub-classes
override it, but I wanted to clean this design a little on this branch.

I did consider making the Object destructor pure virtual, but then all
sub-classes are forced to override it, which may have rule of three/five
ramifications for those sub-classes.