MOS-60 MOS-61 add multi-threaded test with mutex lock
I added a simple test that uses the thread pool to multi-thread
incrementing an integer with the addition that I am using an std::mutex
and an std::lock_guard to guarantee thread safe increments.

I had more complicated tests, but felt it was better to add a simple
example first.

This runs on my Mac, but might fail elsewhere (e.g. Windows).