MOS-56 renamed TimeGrid and redesign to use new Time class
I renamed TimeGrid to TimingInfo, because that's what we're exposing
in the Mosaic 1.2 API and I think the name is more informative.

I then changed the TimingInfo class to use the new Time and Ratio classes.

I also briefly modified the legacy Trace class to use all these new
classes just to get things to compile - note that the Trace class is
pretty broken and useless at the moment.

Finally, I made the appropriate changes to the TimingInfoTest and
added a new one for kicks.