MOS-59 reimplement movie and player with new time classes
I reimplemented the existing API and also added a new method to
access the timing info of a movie.

As the current HDF5 nVista files do not store timing information
I just populated the timing info with placeholder values and added
a TODO.

In Player, I already noticed a potential problem, which is that the
QTimer we're using to control playback only has a temporal resolution
of 1 ms.