MOS-59 integrate time and timing info with player
I made some modifications to our movie player so that it uses the
time and timing info classes.

This allows the current time of the player to be specified as an
absolute time, which is desirable when we have many sources of data
which different timing information.

As part of this, I provided a new method in the timing info class
to convert an absolute time to an index.

Note that the player is still using QTimer, which only allows
millisecond precision, which is undesirable when we have a frame
rate like 30Hz.