MOS-171 add conversion to center of time window
I renamed the current conversion from indices to times to indicate
it gets a start time of a window associated with a sample.

I then add a conversion method that is consistent with spacing info
where the center of sample window is returned.

I also added tests for both and have explicit tests for what happens
when there are no samples.