MOS-246 initial commit
This commit includes enough to compile and pass all tests on a mac.
There are some changes in the application for compilation, but there
has been no effort to make these work properly yet.

The changes include addition of the new Group and DataSet classes
which are use by the project to keep track of data.