MOS-246 rename workspace model and simplify it
Most of the changes here are related to simplifying the project
model. Previously the application effectively had three trees:

1. The QAbtractItemModel
2. The TreeItem tree
3. The Project tree

The structure of (2) and (3) was pretty close to identical. (2) had
some extra functionality related to (1) that made the application
work.

My solution was to remove (2), add a little extra functionality to (3)
and add some utility functions to (1).

I also removed some methods from (1) that related to modification,
which help for doing efficient insertions/removal of items in the
tree.
My simpler approach is to only modify the Project tree and then
refresh the entire layout.