MOS-273 make project store absolute file names for data sets
I created a test to create a synthetic project that contains a
group to test that the visualization can handle groups.
While doing this I ran into a problem because we store relative
file names as they are given, which means that the file name may
be dependent on where the application that creates it is running.

I changed the behavior of the project to always store the absolute
path to the data set file name.
I think this is what we want to do at run time, and we may want to
serialize file names that are in the same directory sub-tree as the
project file name, relative to the project file name (so that if
the whole sub-tree is moved, the relative paths are still accurate).
I created MOS-308 for this feature.